fix(captcha): redis-backed input, ownership check, schema fixes
Address review findings on PR #94 (M1 captcha pipeline).

- Per-press input moves to Redis (`cap:input:{chat}:{user}`,
  TTL = challenge lifetime). The previous handler reconstructed
  input from the caption mask and lost the actual digits, sending
  "0004" instead of "1234" to solve() — captcha was unsolvable.
- Callback handlers reject presses from non-target users via a
  Redis meta lookup (`cap:meta:{chat}:{message}`) before acking,
  with a "this isn't your captcha" toast. Previously a stranger
  could wipe a target's challenge by mashing buttons.
- `is_verified` cache (`cap:verified:{chat}:{user}`, 7d TTL)
  short-circuits the join hot path for returning users; populated
  by both solve() success and /verify.
- Migrate telegram_message_id and moderation_actions.message_id
  from BIGINT to INTEGER (matches teloxide MessageId i32); drop
  the lossy `mid as i32` cast in the expiry job.
- Drop the no-op unban_chat_member before kick_chat_member —
  unban does not clear restrictions; the comment was wrong.
- Drop the redundant early verified_users check in solve() — race
  resolution lives inside the FOR UPDATE branch already.
- Remove dead allowed_updates() — teloxide 0.13's Dispatcher
  auto-introspects allowed updates from the handler tree.
- Wire set_my_commands so /help, /status, /verify appear in
  Telegram's command menu.
- WrongLeft now clears the input buffer + caption mask so the
  user can retry without backspacing four times.
- Drop dead `let _ = half_h;` shadow in the renderer.

+++ server/src/models/captcha_challenge.rs
@@ -11,7 +11,7 @@ pub struct CaptchaChallenge {
     pub user_id: i64,
     pub solution: String,
     pub attempts_left: i16,
-    pub telegram_message_id: Option<i64>,
+    pub telegram_message_id: Option<i32>,
     pub expires_at: DateTime<Utc>,
     pub created_at: DateTime<Utc>,
 }
